feat/OT151-33 Endpoint para listar Usuarios
