+++ app/controllers/api/v1/authentication_controller.rb
@@ -12,7 +12,7 @@ module Api
           token = JsonWebToken.encode(user_id: @user.id)
           time = Time.zone.now + 24.hours.to_i
           render json: { token: token, exp: time.strftime('%m-%d-%Y %H:%M'),
-                         username: @user.username }, status: :ok
+                         user: @user.email }, status: :ok
         else
           render json: { error: 'unauthorized' }, status: :unauthorized
         end
