+++ app/controllers/api/v1/users_controller.rb
@@ -5,6 +5,12 @@ module Api
     class UsersController < ApplicationController
       before_action :authorize_request, except: :create
       before_action :find_user, except: %i[create]
+      before_action :admin, only: %i[index]
+
+      def index
+        @users = User.all
+        render json: @users, status: :ok
+      end
 
       def create
         @user = User.new(user_params)
@@ -18,13 +24,13 @@ module Api
       private
 
       def find_user
-        @user = User.find_by!(username: params[:_username])
+        @user = User.find(params[:id])
       rescue ActiveRecord::RecordNotFound
         render json: { errors: 'User not found' }, status: :not_found
       end
 
       def user_params
-        params.permit(:email, :password, :first_name, :last_name)
+        params.permit(:email, :password, :first_name, :last_name, :role_id)
       end
     end
   end
