+++ config/application.rb
@@ -24,6 +24,8 @@ module OT151Server
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.1
 
+    config.autoload_paths << Rails.root.join('lib')
+
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
