+++ config/routes.rb
@@ -3,6 +3,7 @@
 Rails.application.routes.draw do
   namespace :api, defaults: { format: 'json' } do
     namespace :v1 do
+      resources :users
       post 'auth/register', to: 'users#create'
       post '/auth/login', to: 'authentication#login'
       get '/*a', to: 'application#not_found'
